`field` accepts `value_macro` and `section` accepts `caller`
The `value_macro` keyword-argument on `field` only has an effect on static rendering and allows the caller to parameterize how the value of the field is displayed by passing a macro that will be used to render the field's value.

If a form template wants to use this new argument, it will not be able to use the `section` macro unless it can be configured via the `caller` mechanism in Jinja2. So this commit also adds that functionality to both static and dynamic macros.